Move helper builder function
