+++ Sources/Primer3DS/Classes/Primer3DS.swift
@@ -209,25 +209,5 @@ extension Primer3DS: ChallengeStatusReceiver {
     
 }
 
-extension Transaction {
-    func buildThreeDSecureAuthData() throws -> Primer3DSSDKGeneratedAuthData {
-        let transactionParameters = try self.getAuthenticationRequestParameters()
-        let sdkAppId = transactionParameters.getSDKAppID()
-        let sdkTransactionId = transactionParameters.getSDKTransactionId()
-        let sdkMaxTimeout = 10
-        let sdkEncData = transactionParameters.getDeviceData()
-        let sdkEphemeralKey = transactionParameters.getSDKEphemeralPublicKey()
-        let sdkReferenceNumber = transactionParameters.getSDKReferenceNumber()
-        
-        return SDKAuthData(
-            sdkAppId: sdkAppId,
-            sdkTransactionId: sdkTransactionId,
-            sdkTimeout: sdkMaxTimeout,
-            sdkEncData: sdkEncData,
-            sdkEphemPubKey: sdkEphemeralKey,
-            sdkReferenceNumber: sdkReferenceNumber)
-    }
-}
-
 #endif
 #endif
